- create S3 destination class - extract common file operations to util class - refactor argument parsing both from the command line and into the operator classes. simplified parsing coming into the process, and a preference for passing single objects instead of fields in order to allow for a consistent interface - implement actual 'fully functional' entrypoint to the program in the index file.

+++ src/destination/FileSystemDestination.ts
@@ -1,46 +1,46 @@
 import path from 'path'
 import fs from 'fs'
-import crypto from 'crypto'
 import FileMeta from "../model/FileMeta.js";
+import FileUtils from "../FileUtils.js";
 
 export default class FileSystemDestination {
     basepath: string
     existingHashes: string[] | null
 
-    constructor(basepath: string) {
-        this.basepath = path.resolve(process.cwd(), basepath)
+    constructor(destinationargs: any) {
+        if(!destinationargs){
+            throw Error('Missing or undefined destinationargs found during initialization of FileSystemDestination')
+        }
+        if(!destinationargs.path){
+            throw Error('Missing required argument for destination initialization :: ' + 'destination.path')
+        }
+        this.basepath = path.resolve(process.cwd(), destinationargs.path)
         this.existingHashes = null
     }
 
 
     public async saveMedia(filemeta: FileMeta, rawfilepromise: Promise<Buffer>): Promise<void> {
-        // if(await this.doesFileAlreadyExist(filemeta)){
-        //     return Promise.resolve()
-        // }
+        if(await this.doesFileAlreadyExist(filemeta)){
+            return Promise.resolve()
+        }
 
-        const newpath = path.resolve(this.basepath, this.getUniqueName(filemeta.name))
+        const newpath = path.resolve(this.basepath, FileUtils.getUniqueName(filemeta.name))
         return rawfilepromise
             .then(raw => fs.promises.writeFile(newpath, raw))
-            // .then(_ => this.existingHashes?.push(filemeta.md5))
+            .then(_ => this.existingHashes?.push(filemeta.md5))
             .then(() => Promise.resolve())
     }
 
-    //TODO: shift this into some util class.
-    private getUniqueName(filename: string): string {
-        const shortname = filename.substring(0, filename.lastIndexOf('.'))
-        const timestamp = new Date().getTime()
-        return shortname + '_' + timestamp + '.' + filename.split('.').pop()
-    }
-
-    private async getExistingMeta(): Promise<FileMeta[]> {
+    public async getExistingMeta(): Promise<FileMeta[]> {
         const dirlist = await fs.promises.readdir(this.basepath, {withFileTypes: true})
         const meta = dirlist.filter(x => x.isFile).map(async x => {
             const uri = path.resolve(this.basepath, x.name)
             const rawfile = await fs.promises.readFile(uri)
-            return new FileMeta(path.basename(uri), uri, path.extname(uri), this.getMd5(rawfile))
+            return new FileMeta(path.basename(uri), uri, path.extname(uri), FileUtils.getMd5(rawfile))
         })
-
-        return Promise.all(meta)
+        const metaresult = await Promise.all(meta)
+        this.existingHashes = metaresult.map(x => x.md5)
+        return metaresult
     }
 
     public async doesFileAlreadyExist(filemeta: FileMeta): Promise<boolean> {
@@ -48,10 +48,5 @@ export default class FileSystemDestination {
             this.existingHashes = await this.getExistingMeta().then(x => x.map(z => z.md5))
         }
         return Promise.resolve((this.existingHashes as string[]).includes(filemeta.md5))
-    }
-
-    //TODO: merge this and the one in the fs source
-    private getMd5(rawfile: Buffer): string {
-        return crypto.createHash('md5').update(rawfile).digest('hex')
     }
 }
